refactor(monitor): centralize model configuration to eliminate duplication
- Create model-config.mjs with unified model settings
- Remove duplicate PRICING definitions from usage-calculator.mjs
- Remove duplicate CONTEXT_WINDOWS from context-tracker.mjs
- Remove duplicate model info from session-cache.mjs
- Maintain backward compatibility with re-exports
- All tests passing (196/196)
