+++ src/monitor/usage-calculator.mjs
@@ -1,48 +1,7 @@
-// モデル別の料金設定（USD per 1M tokens）
-export const PRICING = {
-  'claude-3-opus-20241022': {
-    input: 15.00,
-    output: 75.00,
-    name: 'Claude 3 Opus'
-  },
-  'claude-opus-4-20250514': {
-    input: 15.00,
-    output: 75.00,
-    name: 'Claude Opus 4'
-  },
-  'claude-opus-4-1-20250805': {
-    input: 15.00,
-    output: 75.00,
-    name: 'Claude Opus 4.1'
-  },
-  'claude-sonnet-4-20250514': {
-    input: 2.25,
-    output: 11.25,
-    name: 'Claude Sonnet 4'
-  },
-  'claude-3-5-sonnet-20241022': {
-    input: 3.00,
-    output: 15.00,
-    name: 'Claude 3.5 Sonnet'
-  },
-  'claude-3-5-haiku-20241022': {
-    input: 1.00,
-    output: 5.00,
-    name: 'Claude 3.5 Haiku'
-  },
-  'claude-3-haiku-20240307': {
-    input: 0.25,
-    output: 1.25,
-    name: 'Claude 3 Haiku'
-  }
-};
+import { PRICING, getModelName as getModelNameFromConfig, getModelPricing } from './model-config.mjs';
 
-// デフォルト料金（未知のモデル用）
-const DEFAULT_PRICING = {
-  input: 3.00,
-  output: 15.00,
-  name: 'Unknown Model!'
-};
+// Re-export PRICING for backward compatibility
+export { PRICING };
 
 export class UsageCalculator {
   constructor() {
@@ -62,7 +21,7 @@ export class UsageCalculator {
       };
     }
     
-    const pricing = this.pricing[model] || DEFAULT_PRICING;
+    const pricing = getModelPricing(model);
     
     // Convert to numbers, treat invalid values as 0
     const inputTokens = Number(usage.input_tokens) || 0;
@@ -143,8 +102,7 @@ export class UsageCalculator {
   }
 
   getModelName(model) {
-    const info = this.pricing[model];
-    return info ? info.name : 'Unknown Model';
+    return getModelNameFromConfig(model);
   }
 
   estimateRemainingTurns(currentTokens, contextWindow, averageTokensPerTurn) {
